+++ src/components/Quests/DeleteQuest/DeleteQuest.js
@@ -14,9 +14,11 @@ import {
   Center,
   FormControl,
   FormErrorMessage,
+  Tooltip,
 } from "@chakra-ui/react";
 import { DeleteIcon } from "@chakra-ui/icons";
 import FirestoreContext from "../../../store/firestore-context";
+import { auth } from "../../../Firebase";
 
 const DeleteQuest = (props) => {
   const [questNameInput, setQuestNameInput] = useState("");
@@ -30,7 +32,10 @@ const DeleteQuest = (props) => {
 
   const deleteHandler = (event) => {
     event.preventDefault();
-    if (questNameInput === props.questName) {
+    if (
+      questNameInput === props.questName &&
+      props.userId === auth.currentUser.uid
+    ) {
       try {
         fireCtx.DeleteQuestItemInDB(props.questId);
         setQuestDeleteError();
@@ -43,9 +48,16 @@ const DeleteQuest = (props) => {
 
   return (
     <Box>
-      <Button onClick={onOpen}>
-        <DeleteIcon color="red" />
-      </Button>
+      <Box>
+        <Tooltip label="Delete">
+          <DeleteIcon
+            _hover={{ cursor: "pointer" }}
+            onClick={onOpen}
+            color="red"
+            margin="5px"
+          />
+        </Tooltip>
+      </Box>
       <Modal isOpen={isOpen} onClose={onClose}>
         <ModalOverlay />
         <Center>
@@ -56,6 +68,7 @@ const DeleteQuest = (props) => {
             borderRadius="6px"
             width="300px"
             height="100x"
+            margin="30vh"
           >
             <ModalHeader>Delete {props.questName}</ModalHeader>
             <ModalCloseButton />
@@ -70,15 +83,22 @@ const DeleteQuest = (props) => {
                   ) : null}
                 </FormControl>
                 <FormLabel htmlFor="questNameCheck">
-                  Please enter the quest name to delete news
+                  Please enter the quest name to delete quest
                 </FormLabel>
                 <Input
                   type="text"
                   value={questNameInput}
                   onChange={questNameInputHandler}
+                  backgroundColor="#ECF8F8"
                 />
-                <Button type="submit">
-                  <DeleteIcon color="red" />
+                <Button marginLeft="80%" background="none" type="submit">
+                  <Tooltip label="Delete">
+                    <DeleteIcon
+                      _hover={{ cursor: "pointer" }}
+                      onClick={onOpen}
+                      color="red"
+                    />
+                  </Tooltip>
                 </Button>
               </form>
             </ModalBody>
